requirements.txt
pandas
numpy
scikit-learn
boto3
awscli 
pyyaml
matplotlib 
seaborn
imbalanced-learn
imblearn
lightgbm
mlflow